fix: reduced the bundle size
