+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@standard-community/standard-json",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "type": "module",
   "license": "MIT",
   "main": "dist/index.cjs",
@@ -44,15 +44,29 @@
     }
   },
   "peerDependencies": {
-    "@valibot/to-json-schema": "^1.0.0-rc.0",
-    "effect": "^3.13.10",
-    "zod-to-json-schema": "^3.24.1"
+    "@standard-schema/spec": "^1.0.0",
+    "@types/json-schema": "^7.0.15",
+    "@valibot/to-json-schema": "^1.3.0",
+    "arktype": "^2.1.20",
+    "effect": "^3.16.8",
+    "valibot": "^1.1.0",
+    "zod": "^3.25.67",
+    "zod-to-json-schema": "^3.24.5"
   },
   "peerDependenciesMeta": {
     "@valibot/to-json-schema": {
       "optional": true
     },
+    "arktype": {
+      "optional": true
+    },
     "effect": {
+      "optional": true
+    },
+    "valibot": {
+      "optional": true
+    },
+    "zod": {
       "optional": true
     },
     "zod-to-json-schema": {
@@ -61,14 +75,9 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.0.4",
-    "@standard-schema/spec": "^1.0.0",
-    "@types/json-schema": "^7.0.15",
-    "arktype": "^2.0.4",
     "pkgroll": "^2.5.1",
     "typescript": "^5.8.3",
-    "valibot": "^1.0.0-rc.4",
-    "vitest": "^3.2.4",
-    "zod": "^3.24.2"
+    "vitest": "^3.2.4"
   },
-  "packageManager": "pnpm@8.14.0+sha256.9cebf61abd83f68177b29484da72da9751390eaad46dfc3072d266bfbb1ba7bf"
+  "packageManager": "pnpm@10.0.0"
 }
