+++ src/vendors/arktype.ts
@@ -1,7 +1,9 @@
 import type { Type } from "arktype";
 import type { JSONSchema7 } from "json-schema";
-import type { ToJsonSchemaFn } from "./index.js";
+import type { ToJsonSchemaFn } from "./utils.js";
 
-export const getToJsonSchemaFn =
+const getToJsonSchemaFn =
   async (): Promise<ToJsonSchemaFn> => (schema) =>
     (schema as Type).toJsonSchema() as JSONSchema7;
+
+export default getToJsonSchemaFn;
