+++ src/vendors/valibot.ts
@@ -1,7 +1,7 @@
 import type { BaseIssue, BaseSchema } from "valibot";
-import { tryImport, type ToJsonSchemaFn } from "./index.js";
+import { tryImport, type ToJsonSchemaFn } from "./utils.js";
 
-export const getToJsonSchemaFn = async (): Promise<ToJsonSchemaFn> => {
+const getToJsonSchemaFn = async (): Promise<ToJsonSchemaFn> => {
   const { toJsonSchema } = await tryImport(
     import("@valibot/to-json-schema"),
     "@valibot/to-json-schema",
@@ -12,3 +12,5 @@ export const getToJsonSchemaFn = async (): Promise<ToJsonSchemaFn> => {
       options as Record<string, unknown>,
     );
 };
+
+export default getToJsonSchemaFn;
