+++ src/vendors/zod.ts
@@ -1,8 +1,8 @@
 import type { JSONSchema7 } from "json-schema";
 import type { ZodSchema } from "zod";
-import { tryImport, type ToJsonSchemaFn } from "./index.js";
+import { tryImport, type ToJsonSchemaFn } from "./utils.js";
 
-export const getToJsonSchemaFn = async (): Promise<ToJsonSchemaFn> => {
+const getToJsonSchemaFn = async (): Promise<ToJsonSchemaFn> => {
   const { zodToJsonSchema } = await tryImport(
     import("zod-to-json-schema"),
     "zod-to-json-schema",
@@ -13,3 +13,5 @@ export const getToJsonSchemaFn = async (): Promise<ToJsonSchemaFn> => {
       options as Record<string, unknown>,
     ) as JSONSchema7;
 };
+
+export default getToJsonSchemaFn;
